refactor: place Carquet adapter headers in extension includes
