+++ extension/parquet/src/carquet/sniffer.cc
@@ -9,7 +9,7 @@
 #include <string>
 
 #include "carquet/schema_converter.h"
-#include "input_adapter.h"
+#include "carquet/input_adapter.h"
 
 namespace neug::parquet {
